# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      id:
        type: integer
      school:
        $ref: '#/definitions/models.School'
      school_id:
        type: integer
      student:
        $ref: '#/definitions/models.Student'
      student_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Class:
    properties:
      academic_year:
        type: string
      created_at:
        type: string
      homeroom:
        type: string
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      school:
        $ref: '#/definitions/models.School'
      school_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.CreateBook:
    properties:
      author:
        type: string
      school_id:
        type: integer
      student_id:
        type: integer
      title:
        type: string
    required:
    - author
    - title
    type: object
  models.CreateGrade:
    properties:
      academic_year:
        type: string
      book_id:
        type: integer
      knowledge_score:
        maximum: 100
        minimum: 0
        type: number
      notes:
        type: string
      school_id:
        type: integer
      semester:
        maximum: 2
        minimum: 1
        type: integer
      skill_score:
        maximum: 100
        minimum: 0
        type: number
      student_id:
        type: integer
    required:
    - academic_year
    - book_id
    - knowledge_score
    - semester
    - skill_score
    - student_id
    type: object
...
  models.Grade:
    properties:
      academic_year:
        type: string
      book:
        $ref: '#/definitions/models.Book'
      book_id:
        type: integer
      created_at:
        type: string
      final_score:
        type: number
      id:
        type: integer
      knowledge_score:
        type: number
      notes:
        type: string
      school:
        $ref: '#/definitions/models.School'
      school_id:
        type: integer
      semester:
        type: integer
      skill_score:
        type: number
      student:
        $ref: '#/definitions/models.Student'
      student_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.School:
    properties:
      address:
        type: string
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Student:
    properties:
      class:
        $ref: '#/definitions/models.Class'
      class_id:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      school:
        $ref: '#/definitions/models.School'
      school_id:
        type: integer
      type:
        enum:
        - junior
        - senior
        example: junior
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateBook:
    properties:
      author:
        type: string
      school_id:
        type: integer
      student_id:
        type: integer
      title:
        type: string
    type: object
  models.UpdateGrade:
    properties:
      knowledge_score:
        maximum: 100
        minimum: 0
        type: number
      notes:
        type: string
      school_id:
        type: integer
      skill_score:
        maximum: 100
        minimum: 0
        type: number
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Raport API
  version: "1.0"
paths:
...
  /books:
    get:
      description: Get a list of all books with optional pagination
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of books
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all books with pagination
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create a new book with the given input data
      parameters:
      - description: Create book object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateBook'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      description: Delete the book with the given ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted book
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a book by ID
      tags:
      - books
    get:
      description: Get details of a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved book
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Book not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Find a book by ID
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update the book details for the given ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Update book object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBook'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a book by ID
      tags:
      - books
...
  /grades:
    get:
      description: 'Get grade list with optional filters: student_id, semester, academic_year'
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      - description: Student ID
        in: query
        name: student_id
        type: integer
      - description: Semester
        in: query
        name: semester
        type: integer
      - description: Academic year
        in: query
        name: academic_year
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of grades
          schema:
            items:
              $ref: '#/definitions/models.Grade'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get grades
      tags:
      - grades
    post:
      consumes:
      - application/json
      description: Create grade for a student and subject(book)
      parameters:
      - description: Create grade object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateGrade'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created grade
          schema:
            $ref: '#/definitions/models.Grade'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Create a grade
      tags:
      - grades
  /grades/{id}:
    delete:
      parameters:
      - description: Grade ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted grade
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete grade by ID
      tags:
      - grades
    put:
      consumes:
      - application/json
      parameters:
      - description: Grade ID
        in: path
        name: id
        required: true
        type: string
      - description: Update grade object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateGrade'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated grade
          schema:
            $ref: '#/definitions/models.Grade'
      security:
      - ApiKeyAuth: []
      summary: Update grade by ID
      tags:
      - grades
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  JwtAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"